Move server role helper to separate file
